rename benchmark module -> model_runner
